# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Predix Agent API",
    "description": "REST API for AI agents and bots to trade on the Predix prediction market. 1-minute BTC price rounds on Stacks testnet with zero gas fees.",
    "version": "0.1.0"
  },
  "servers": [
    { "url": "https://bitpredix.vercel.app", "description": "Production (testnet)" }
  ],
  "paths": {
...
    "/api/agent/register": {
      "post": {
        "summary": "Register agent and get API key",
        "description": "Prove wallet ownership via Stacks signature to receive an API key (pk_live_...). Key is shown once.",
        "operationId": "registerAgent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["wallet", "signature", "message"],
                "properties": {
                  "wallet": { "type": "string", "description": "Stacks address (ST...)" },
                  "signature": { "type": "string", "description": "Signed message hex" },
                  "message": { "type": "string", "description": "Predix Agent Registration {timestamp}" },
                  "name": { "type": "string", "maxLength": 32 },
                  "description": { "type": "string", "maxLength": 200 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Registration success with API key" }
        }
      }
    },
...
    "/api/agent/webhooks": {
      "post": {
        "summary": "Create webhook",
        "description": "Subscribe to events (round.resolved, bet.confirmed, etc). Max 5 per agent.",
        "operationId": "createWebhook",
        "security": [{ "AgentApiKey": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["url", "events"],
                "properties": {
                  "url": { "type": "string", "description": "HTTPS webhook URL" },
                  "events": { "type": "array", "items": { "type": "string" } },
                  "secret": { "type": "string", "description": "Optional HMAC secret" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Webhook created with secret" } }
      },
      "get": {
        "summary": "List webhooks",
        "operationId": "listWebhooks",
        "security": [{ "AgentApiKey": [] }],
        "responses": { "200": { "description": "List of agent webhooks" } }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "AgentApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Predix-Key",
        "description": "Agent API key (pk_live_...). Register at POST /api/agent/register."
      }
    },
...
  }
}